requirements.txt
pandas
matplotlib
torch
Pillow
transformers
datasets
timm
accelerate==1.6.0
wandb
torchmetrics
scipy
pycocotools